# modules/module-automation-video-api/openapi/v1/video.yaml
openapi: 3.1.0
info:
  title: Automation Video API
  version: 1.0.0
  description: Versioned, authenticated API contract for the video module.
servers:
  - url: /
paths:
...
  /api/v1/automation/video/{id}/transition:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    post:
      operationId: automation.video.transition
      summary: Transition the resource through an allowed lifecycle state.
      security:
        - sanctum: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [draft, active, paused, completed, failed, cancelled]
      responses:
        '200':
          $ref: '#/components/responses/Resource'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      description: Opaque resource identifier.
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  schemas:
...
    Resource:
      type: object
      required:
        - id
        - type
        - attributes
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          const: automation-video
        attributes:
          type: object
          additionalProperties: true
    ResourceResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/Resource'
...
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
  responses:
    Resource:
      description: Authorized resource response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceResponse'
...
    Problem:
      description: RFC 9457 Problem Details response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'